+++ src/graphic/render/gl_surface_screen.h
@@ -28,7 +28,7 @@
  */
 class GLSurfaceScreen : public GLSurface {
 public:
-	GLSurfaceScreen(uint32_t w, uint32_t h);
+	GLSurfaceScreen(uint16_t w, uint16_t h);
 
 	/// Interface implementations
 	virtual void lock(LockMode);
